% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{available_releases}
\alias{available_releases}
\title{Retrieve a listing of releases}
\usage{
available_releases(...)
}
\arguments{
\item{...}{other arguments for \code{\link[rfishbase]{available_releases}}}
}
\value{
see \code{\link[rfishbase]{available_releases}}
}
\description{
This is a wrapper around \code{\link[rfishbase]{available_releases}} which
allows us to impoirt the package.  We don't need this, but we do need the
import if we are to use \code{getFromNamespace()} to get functions by name.
}
