% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_entrez_key}
\alias{get_entrez_key}
\title{Retrieve the user's entrez_key from ~/.entrez_key}
\usage{
get_entrez_key(filename = "~/.entrez_key")
}
\arguments{
\item{filename}{chr the name of the entrez key filename}
}
\value{
the entrez key
}
\description{
Retrieve the user's entrez_key from ~/.entrez_key
}
