% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{label_species_list}
\alias{label_species_list}
\title{Label a species table by user specified groups.  Each row is assigned to at
least one group or the default group.}
\usage{
label_species_list(
  x = read_species_list(),
  variable = "Class",
  groups = list(fish = c("Actinopterygii", "Elasmobranchii")),
  default = NA_character_
)
}
\arguments{
\item{x}{table of species}

\item{variable}{char name of the column to filter}

\item{groups}{named list, one or more character vectors of values belong to the group(s)}

\item{default}{chr, the default grouping for ungroup rows}
}
\value{
the input with new column of group.  Ungrouped rows are assigned the default.
}
\description{
Label a species table by user specified groups.  Each row is assigned to at
least one group or the default group.
}
