% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_configuration}
\alias{read_configuration}
\title{Read a configuration}
\usage{
read_configuration(filename, version = NULL, root_path = ".", template = FALSE)
}
\arguments{
\item{filename}{chr, the full name of the file, ignored if version is not NULL}

\item{version}{chr, a segemented version ID such as 'v1.234' or NULL}

\item{root_path}{chr, the root path}

\item{template}{logical, if TRUE read the provided example}
}
\value{
a configuration
}
\description{
Read a configuration
}
