% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{read_species_list}
\alias{read_species_list}
\title{Read a species list}
\usage{
read_species_list(
  filename,
  keep = c("ScientificName_accepted", "Phylum/Division", "Class", "Order", "Family")
)
}
\arguments{
\item{filename}{chr, the name of the file (with extension and path)}

\item{keep}{chr, one or more column names to keep.  "all" to keep them all.}
}
\description{
Read a species list
}
