% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refdbtools-package.R
\docType{package}
\name{refdbtools}
\alias{refdbtools}
\alias{refdbtools-package}
\title{refdbtools package}
\description{
R language package to assist in making eDNA reference databases .
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bigelowlab/refdbtools}
}

}
