% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi.R
\name{search_target_species}
\alias{search_target_species}
\title{Search of target species at NCBI}
\usage{
search_target_species(cfg, entrez_key = get_entrez_key())
}
\arguments{
\item{cfg}{configuration list}

\item{entrez_key}{character, see \code{\link{get_entrez_key}}}
}
\value{
list of voucher and non-voucher search results
}
\description{
Search of target species at NCBI
}
