% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\name{select_target_species}
\alias{select_target_species}
\title{A function narrow the scope of species to include in the database}
\usage{
select_target_species(cfg, save_config = TRUE)
}
\arguments{
\item{cfg}{a configuration list}

\item{save_config}{logical, if TRUE save the provided config in the destination
directory}
}
\value{
a table of target species
}
\description{
User provides a configuration file with fishbase and/or sealifebase search
specifications. User may also provide an optional "preselection" to narrow
the search range.
}
