% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{version_path}
\alias{version_path}
\title{Compose a version path}
\usage{
version_path(version = "v0.000", root_path = ".")
}
\arguments{
\item{version}{chr, a segemented version ID such as 'v1.234'}

\item{root_path}{chr, the root path}
}
\value{
a path description
}
\description{
Compose a version path
}
