% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_configuration}
\alias{write_configuration}
\title{Read a configuration}
\usage{
write_configuration(cfg, root_path = NULL)
}
\arguments{
\item{cfg}{list, configuration list}

\item{root_path}{chr, the root path or NULL in which case the root path is taken
from the config it self}
}
\value{
the input configuration
}
\description{
Read a configuration
}
